# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Bank Cards Api
  description: API системы управления банковскими картами
  contact:
    name: Nazarov Oleg
    email: noleg867@gmail.com
  version: 1.0.0
servers:
- url: http://localhost:8080
  description: Generated server url
...
paths:
...
  /api/auth/sign-up:
    post:
      tags:
      - Контроллер для регистрации/аутентификации.
      summary: Регистрация пользователя.
      description: Позволяет зарегистрироваться новому пользователю.
      operationId: signUp
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignUp"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
  /api/auth/sign-in:
    post:
      tags:
      - Контроллер для регистрации/аутентификации.
      summary: Аутентификация пользователя.
      description: Позволяет повторно войти уже зарегистрированному пользователю.
      operationId: signIn
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignIn"
        required: true
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JwtResponse"
  /api/admin/users/{userId}:
    delete:
      tags:
      - Контроллер для администратора.
      summary: Удаление пользователя.
      description: Позволяет удалить любого пользователя по id (кроме администраторов).
      operationId: deleteUser
      parameters:
      - name: userId
        in: path
        description: Идентификатор пользователя
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      responses:
        "200":
          description: OK
      security:
      - JWT: []
    patch:
      tags:
      - Контроллер для администратора.
      summary: Обновление роли.
      description: Позволяет обновить роль пользователя.
      operationId: updateUserRole
      parameters:
      - name: userId
        in: path
        description: Идентификатор пользователя
        required: true
        schema:
          type: integer
          format: int64
          minimum: 1
      - name: role
        in: query
        description: Роль пользователя
        required: true
        schema:
          type: string
          enum:
          - ROLE_USER
          - ROLE_ADMIN
      responses:
        "200":
          description: OK
      security:
      - JWT: []
...
  /api/admin/users:
    get:
      tags:
      - Контроллер для администратора.
      summary: Получение всех пользователей.
      description: Позволяет получить всех пользователей системы.
      operationId: getAllUsers
      parameters:
      - name: pageNumber
        in: query
        description: Номер страницы
        required: true
        schema:
          type: integer
          format: int32
          minimum: 0
      - name: pageSize
        in: query
        description: Размер страницы
        required: true
        schema:
          type: integer
          format: int32
          maximum: 10
          minimum: 1
      - name: sort
        in: query
        description: Сортировка
        required: true
        schema:
          type: string
          enum:
          - ID_ASC
          - ID_DESC
          - LAST_NAME_ASC
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/UserDto"
      security:
      - JWT: []
components:
  schemas:
...
    SignUp:
      type: object
      description: Регистрация пользователя
      properties:
        email:
          type: string
          description: Уникальный email пользователя
          example: user123@gmail.com
          maxLength: 50
          minLength: 5
        password:
          type: string
          description: Пароль пользователя
          example: user_password123
        firstName:
          type: string
          description: Имя пользователя
          example: Иван
          maxLength: 50
          minLength: 5
        lastName:
          type: string
          description: Фамилия пользователя
          example: Иванов
          maxLength: 50
          minLength: 5
        patronymic:
          type: string
          description: Отчество пользователя
          example: Иванович
          maxLength: 50
          minLength: 0
      required:
      - email
      - firstName
      - lastName
      - password
    SignIn:
      type: object
      description: Аутентификация пользователя
      properties:
        email:
          type: string
          description: Уникальный email пользователя
          example: user123@gmail.com
          maxLength: 50
          minLength: 5
        password:
          type: string
          description: Пароль пользователя
          example: user_password123
      required:
      - email
      - password
    JwtResponse:
      type: object
      description: Ответ при аутентификации
      properties:
        token:
          type: string
          description: Jwt токен
...
    UserDto:
      type: object
      description: Пользователь
      properties:
        id:
          type: integer
          format: int64
          description: Id пользователя
          example: 1
          readOnly: true
        firstName:
          type: string
          description: Имя пользователя
          example: Иван
        lastName:
          type: string
          description: Фамилия пользователя
          example: Иванов
        patronymic:
          type: string
          description: Отчество пользователя
          example: Иванович
        email:
          type: string
          description: Email пользователя
          example: example@gmail.com
        role:
          type: string
          description: Роль пользователя
          enum:
          - ROLE_USER
          - ROLE_ADMIN
          example: ROLE_USER
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT